+++ RTInAWeekend/inc/material.h
@@ -104,7 +104,7 @@ public:
 private:
 	//Approximation for varying reflectivity with angle.
 	// Uses Schlick's Approximation for reflectance.
-	static double reflectance(float cosine, float ref_idx) {
+	static double reflectance(double cosine, double ref_idx) {
 		auto r0 = (1 - ref_idx) / (1 + ref_idx);
 		r0 *= r0;
 		return r0 + (1 - r0) * pow(1 - cosine, 5);
